Remove Eto
Unfortunately, Eto is.. not the best of frameworks in a lot of ways. It seems like it would work for a more traditional GUI app which isn't so dependent on having a performant renderer for more specialized things. The native theming (at least on Winforms) feels old and the Wpf integration doesn't cooperate very well.