Fix sampling issues
Fillers and rooms shouldn't sometimes have pixels of tear between them anymore. Solved for rooms by switching back to PointClamp. Solved for fillers (and all rectangles) by changing the pixel texture to be the middle of a 3x3 rectangle on the spritesheet, so when the GPU samples from an adjacent pixel it is not transparent (this is what caused the tearing in the first place as far as I'm aware.)